fix datagouv components wrong config after 404
